修改 navigation bar 样式，封装警告框弹出方式
