+++ SwiftGG/Application/AppDelegate.swift
@@ -25,7 +25,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         let loginController = SGLoginViewController()
         let navigationViewController = UINavigationController(rootViewController: loginController)
         window!.rootViewController = navigationViewController
-        setClearNavigationBar()
         
         window!.backgroundColor = UIColor.whiteColor()
         window!.makeKeyAndVisible()
@@ -38,13 +37,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         }
         
         return true
-    }
-    
-    func setClearNavigationBar() {
-        let navigationController = window?.rootViewController as! UINavigationController
-        navigationController.navigationBar.tintColor = UIColor.clearColor()
-        navigationController.navigationBar.setBackgroundImage(UIImage.st_imageWithColor(UIColor.clearColor()), forBarMetrics: .Default)
-        navigationController.navigationBar.clipsToBounds = true
     }
     
     func setDefaultAppearance(){
